requirements.txt
Django
django-mongoengine
six
pymongo[srv]
djangorestframework
django-rest-framework-mongoengine
mysqlclient
Pillow
django-cors-headers
psycopg2
blinker
redis
